% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary-analysis.R
\name{summarize_stats_CSV}
\alias{summarize_stats_CSV}
\title{Summary Statistics for MGDrivE2}
\usage{
summarize_stats_CSV(
  read_dir,
  write_dir = read_dir,
  mean = TRUE,
  quantiles = NULL,
  spn_P,
  tmax,
  dt,
  rem_file = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{read_dir}{Directory to find repetition folders in}

\item{write_dir}{Directory to write output}

\item{mean}{Boolean, calculate mean or not. Default is TRUE}

\item{quantiles}{Vector of quantiles to calculate. Default is NULL}

\item{spn_P}{Places object, see details}

\item{tmax}{The final time to end simulation}

\item{dt}{The time-step at which to return output (\strong{not} the time-step of the sampling algorithm)}

\item{rem_file}{Remove original output? Default is FALSE}

\item{verbose}{Chatty? Default is TRUE}
}
\value{
Writes output to files in write_dir
}
\description{
This function reads in all repetitions for each patch and calculates either
the mean, quantiles, or both. User chooses the quantiles, up to 4 decimal places,
and enters them as a vector. Quantiles are calculated empirically. (order does not matter)  \cr
}
\details{

}
